# api/openapi.yaml
openapi: 3.0.3
info:
  title: 『幻辞』 Genji API
  description: |
    幻辞プロジェクトの語彙データベース API（Go + Gin 実装）。約21万5千語を収録。
    見出し語・読みの完全一致検索、全文検索（FTS5）、ランダム取得を提供する read-only API。
  version: "1.0.0"
  license:
    name: MIT
    url: https://github.com/Iktahana/Genji/blob/main/LICENSE
  contact:
    name: Genji
    url: https://github.com/Iktahana/Genji

servers:
  - url: /
    description: Genji API server

tags:
  - name: lookup
    description: 完全一致検索
  - name: search
    description: 全文検索（FTS5）
  - name: entries
    description: エントリ取得
  - name: system
    description: システム情報

paths:
  /:
    get:
      tags: [system]
      operationId: getApiInfo
      summary: API トップ情報
      description: 辞書のバージョン・収録語数・ビルド日時・主要エンドポイント／ドキュメントへのリンクを返す
      responses:
        "200":
          description: API 情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiInfo"

  /healthz:
    get:
      tags: [system]
      operationId: getHealth
      summary: ヘルスチェック
      description: DB 接続を確認する
      responses:
        "200":
          description: 正常
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: DB 接続不可
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/metadata:
    get:
      tags: [system]
      operationId: getMetadata
      summary: ビルドメタデータ取得
      description: データベースの _metadata テーブル（バージョン・コミット・収録語数等）を返す
      responses:
        "200":
          description: メタデータ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Metadata"

  /v1/entries/{uuid}:
    get:
      tags: [entries]
      operationId: getEntryByUUID
      summary: UUID でエントリ取得
      parameters:
        - name: uuid
          in: path
          required: true
          description: エントリの UUID
          schema:
            type: string
      responses:
        "200":
          description: エントリ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: 見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/lookup/entry:
    get:
      tags: [lookup]
      operationId: lookupByEntry
      summary: 見出し語で検索（完全一致）
      parameters:
        - name: word
          in: query
          required: true
          description: 見出し語（漢字・カタカナ等）
          schema:
            type: string
      responses:
        "200":
          description: 一致したエントリ一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryList"

  /v1/lookup/reading:
    get:
      tags: [lookup]
      operationId: lookupByReading
      summary: 読みで検索（完全一致）
      parameters:
        - name: reading
          in: query
          required: true
          description: 読み（ひらがな・カタカナ）
          schema:
            type: string
      responses:
        "200":
          description: 一致したエントリ一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryList"

  /v1/search/entries:
    get:
      tags: [search]
      operationId: searchEntries
      summary: 見出し語・読みを全文検索
      description: 見出し語または読みを FTS5 で検索する
      parameters:
        - name: q
          in: query
          required: true
          description: 検索クエリ
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: 取得件数の上限
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: 検索結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResultList"

  /v1/search/definitions:
    get:
      tags: [search]
      operationId: searchDefinitions
      summary: 語釈を全文検索
      description: 語釈（gloss）を FTS5 で検索する
      parameters:
        - name: q
          in: query
          required: true
          description: 検索クエリ（英語の語釈にマッチ）
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: 取得件数の上限
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: 検索結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinitionSearchResultList"

  /v1/random:
    get:
      tags: [entries]
      operationId: randomEntries
      summary: ランダムな語を取得
      parameters:
        - name: count
          in: query
          required: false
          description: 取得件数
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: ランダムなエントリ一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryList"

  /v1/sitemap:
    get:
      tags: [sitemap]
      operationId: getSitemap
      summary: 全収録語彙を熱度順にページングして取得
      description: >
        sitemap 生成用に、全収録語彙を熱度（人気度）の降順でページングして返す。
        熱度 = 直近30日のアクセス数 × 2 + 直近365日のアクセス数 × 1。
        Redis 未設定時は freq_rank → 見出し語の安定順で返し、heat は null になる。
      parameters:
        - name: page
          in: query
          required: false
          description: ページ番号（1 始まり）
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          required: false
          description: 1 ページあたりの件数
          schema:
            type: integer
            default: 1000
            minimum: 1
            maximum: 50000
      responses:
        "200":
          description: 熱度順の語彙ページ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitemapPage"

components:
  schemas:
    EndpointInfo:
      type: object
      required: [path, summary]
      properties:
        path:
          type: string
        summary:
          type: string

    ApiInfo:
      type: object
      description: API トップページの基本情報
      required: [name, description, docs, openapi, frontend, endpoints]
      properties:
        name:
          type: string
          example: "『幻辞』 Genji API"
        description:
          type: string
        version:
          type: string
          nullable: true
          description: 辞書データのバージョン（_metadata 由来）
        entry_count:
          type: string
          nullable: true
          description: 収録語彙数
        build_date:
          type: string
          nullable: true
        commit_short:
          type: string
          nullable: true
        docs:
          type: string
          description: API ドキュメント（Redoc）の URL
          example: "/docs"
        openapi:
          type: string
          description: OpenAPI 仕様の URL
          example: "/openapi.yaml"
        frontend:
          type: string
          description: 利用者向け前端サイト
          example: "https://dict.illusions.app"
        endpoints:
          type: array
          description: 主要エンドポイント一覧
          items:
            $ref: "#/components/schemas/EndpointInfo"

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: "entry not found"

    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: "ok"
        cache:
          type: string
          description: キャッシュの状態（enabled / disabled）
          example: "disabled"

    Metadata:
      type: object
      description: _metadata テーブルの内容
      properties:
        version:
          type: string
        commit:
          type: string
        commit_short:
          type: string
        branch:
          type: string
        repository:
          type: string
        build_date:
          type: string
        entry_count:
          type: string

    Reading:
      type: object
      properties:
        primary:
          type: string
          nullable: true
        alternatives:
          type: array
          items:
            type: string
        is_heteronym:
          type: boolean

    Grammar:
      type: object
      properties:
        pos:
          type: array
          items:
            type: string
        ctype:
          type: string
          nullable: true
        inflections:
          nullable: true
          description: 活用情報（任意の JSON）

    Citation:
      type: object
      properties:
        source:
          type: string
          nullable: true
        author:
          type: string
          nullable: true
        note:
          type: string
          nullable: true

    Example:
      type: object
      properties:
        text:
          type: string
        citation:
          $ref: "#/components/schemas/Citation"

    Examples:
      type: object
      properties:
        standard:
          type: array
          items:
            $ref: "#/components/schemas/Example"
        literary:
          type: array
          items:
            $ref: "#/components/schemas/Example"

    SensoryTags:
      type: object
      properties:
        colors:
          type: array
          items:
            type: string
        temperature:
          type: string
          nullable: true
        sounds:
          type: array
          items:
            type: string
        emotions:
          type: array
          items:
            type: string

    Definition:
      type: object
      properties:
        index:
          type: integer
        gloss:
          type: string
          nullable: true
        register:
          type: string
          nullable: true
        nuance:
          type: string
          nullable: true
        scenarios:
          type: array
          items:
            type: string
        sensory_tags:
          $ref: "#/components/schemas/SensoryTags"
        collocations:
          type: array
          items:
            type: string
        examples:
          $ref: "#/components/schemas/Examples"

    Relations:
      type: object
      properties:
        homophones:
          type: array
          items:
            type: string
        synonyms:
          type: array
          items:
            type: string
        antonyms:
          type: array
          items:
            type: string
        related:
          type: array
          items:
            type: string

    EntryMeta:
      type: object
      properties:
        version:
          type: string
        source:
          type: string
        updated_at:
          type: string
        freq_rank:
          type: integer
          nullable: true
          description: "頻度ランク（小さいほど高頻度。hingston/japanese 由来）"
        frequencies:
          type: object
          additionalProperties:
            type: integer
          description: '出典別の総出現回数。例: {"aozora": 1234}'

    Entry:
      type: object
      required: [uuid, entry]
      properties:
        uuid:
          type: string
        entry:
          type: string
        reading:
          $ref: "#/components/schemas/Reading"
        grammar:
          $ref: "#/components/schemas/Grammar"
        definitions:
          type: array
          items:
            $ref: "#/components/schemas/Definition"
        relations:
          $ref: "#/components/schemas/Relations"
        meta:
          $ref: "#/components/schemas/EntryMeta"

    EntryList:
      type: object
      required: [count, entries]
      properties:
        count:
          type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"

    SearchResult:
      type: object
      required: [uuid, entry]
      properties:
        uuid:
          type: string
        entry:
          type: string
        reading_primary:
          type: string
          nullable: true
        pos:
          type: array
          items:
            type: string
        match_highlight:
          type: string
          description: マッチ箇所のハイライト（<b> タグ付き）

    SearchResultList:
      type: object
      required: [count, results]
      properties:
        count:
          type: integer
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"

    DefinitionSearchResult:
      type: object
      required: [uuid, entry]
      properties:
        uuid:
          type: string
        entry:
          type: string
        reading_primary:
          type: string
          nullable: true
        gloss:
          type: string
          nullable: true
        def_index:
          type: integer
        match_highlight:
          type: string
          description: マッチ箇所のハイライト（<b> タグ付き）

    DefinitionSearchResultList:
      type: object
      required: [count, results]
      properties:
        count:
          type: integer
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/DefinitionSearchResult"

    SitemapItem:
      type: object
      required: [uuid, entry]
      properties:
        uuid:
          type: string
        entry:
          type: string
        reading_primary:
          type: string
          nullable: true
        heat:
          type: number
          format: double
          nullable: true
          description: 熱度指数（2×直近30日 + 1×直近365日のアクセス数）。Redis 無効時は null
        updated_at:
          type: string
          nullable: true
          description: エントリの更新日時（sitemap の lastmod 用）

    SitemapPage:
      type: object
      required: [page, page_size, total, total_pages, items]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          description: 全収録語彙数
        total_pages:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/SitemapItem"